refactor(providers): move ProviderRow and provider list logic
